feat: Implement Monthly and Yearly Savings Statistics Repositories
- Added MonthlyTransactionStatisticRepositoryImpl for handling monthly transaction statistics.
- Added YearlySavingsStatisticRepositoryImpl for managing yearly savings statistics.
- Created MongoDB repositories for daily, monthly, and yearly statistics.
- Introduced data models for DailyTransactionStatistic, MonthlySavingsStatistic, and YearlySavingsStatistic.
- Developed REST resources for accessing savings statistics.
- Enhanced TransactionChangesService to process transaction changes and update statistics.
- Added utility classes for generating test data and mock statistics.
- Implemented integration tests for savings statistics resource.
- Updated application configuration for MongoDB and Kafka.

+++ src/main/kotlin/org/balhom/statisticsapi/modules/currencyprofilechanges/domain/props/CurrencyProfileChangeProps.kt
@@ -2,14 +2,15 @@ package org.balhom.statisticsapi.modules.currencyprofilechanges.domain.props
 
 import org.balhom.statisticsapi.common.data.enums.EventChangeTypeEnum
 import org.balhom.statisticsapi.modules.currencyprofilechanges.domain.models.CurrencyProfileSharedUser
+import java.math.BigDecimal
 import java.util.*
 
 data class CurrencyProfileChangeProps(
     val eventChangeTypeEnum: EventChangeTypeEnum,
     val currencyProfileId: UUID,
-    val balance: Double,
-    val monthlyGoal: Double,
-    val yearlyGoal: Double,
+    val balance: BigDecimal,
+    val monthlyGoal: BigDecimal,
+    val yearlyGoal: BigDecimal,
     val ownerId: UUID,
     val sharedUsers: MutableList<CurrencyProfileSharedUser>,
 )
