+++ src/main/kotlin/org/balhom/statisticsapi/modules/currencyprofilechanges/infrastructure/consumers/data/CurrencyProfileChangeEvent.kt
@@ -3,14 +3,15 @@ package org.balhom.statisticsapi.modules.currencyprofilechanges.infrastructure.c
 import org.balhom.statisticsapi.common.data.enums.EventChangeTypeEnum
 import org.balhom.statisticsapi.modules.currencyprofilechanges.domain.models.CurrencyProfileSharedUser
 import org.balhom.statisticsapi.modules.currencyprofilechanges.domain.props.CurrencyProfileChangeProps
+import java.math.BigDecimal
 import java.util.UUID
 
 data class CurrencyProfileChangeEvent(
     val action: String,
     val id: UUID,
-    val balance: Double,
-    val monthlyGoal: Double,
-    val yearlyGoal: Double,
+    val balance: BigDecimal,
+    val monthlyGoal: BigDecimal,
+    val yearlyGoal: BigDecimal,
     val ownerId: UUID,
     val sharedUsers: MutableList<CurrencyProfileSharedUser>,
 ) {
